requirements.txt
# Core Deep Learning Frameworks
torch>=1.10.0
torchvision>=0.11.0

# MMSegmentation and dependencies
mmcv>=2.0.0
mmengine>=0.7.0
mmsegmentation>=1.0.0

# MMagic for Super Resolution
mmagic>=1.0.0

# Image Processing
opencv-python>=4.5.0
scikit-image>=0.19.0
pillow>=9.0.0

# Data Processing
pandas>=1.3.0
openpyxl>=3.0.0
numpy>=1.21.0

# Map Generation
folium>=0.12.0
branca>=0.4.0

# Utility
slidingwindow>=0.0.13

